pyproject.toml
[project]
name = "excalibur-mcp"
version = "0.6.6"
description = "FastMCP server for posting formatted content to X (Twitter) — Tollbooth-monetized, DPYC-native"
authors = [{name = "eXcalibur MCP Contributors"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["mcp", "twitter", "x", "posting", "ai", "claude", "fastmcp", "tollbooth"]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "cryptography>=46.0.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tollbooth-dpyc[nostr]>=0.1.63",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
png = [
    "cairosvg>=2.7.0",
]

[project.scripts]
excalibur-mcp = "excalibur_mcp.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=excalibur_mcp --cov-report=term-missing"
